+++ diagnostic.php
@@ -57,6 +57,12 @@ if ( ! extension_loaded( 'mbstring' ) )
 	$error[] = 'PHP extensions: RosarioSIS relies on the mbstring extension. Please install and activate it.';
 }
 
+// Check for gd extension.
+if ( ! extension_loaded( 'gd' ) )
+{
+	$error[] = 'PHP extensions: RosarioSIS relies on the gd extension. Please install and activate it.';
+}
+
 // Check for xml extension.
 if ( ! extension_loaded( 'xml' ) )
 {